# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "This Blog API",
        "description": "Implementation of blog API",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/v1/article/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Article Details",
                "description": "Display Details of an article",
                "operationId": "Article Details",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Returns when article is not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Article not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "success": {
                                                    "type": "boolean",
                                                    "example": "true"
                                                },
                                                "data": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "example": "50"
                                                        },
                                                        "title": {
                                                            "type": "string",
                                                            "example": "Iure voluptatem et optio sint iusto architecto et dicta"
                                                        },
                                                        "cover_image_url": {
                                                            "type": "string",
                                                            "example": "https://via.placeholder.com/150/09f/fff.png"
                                                        },
                                                        "body": {
                                                            "type": "string",
                                                            "example": "Culpa facere laborum asperiores reiciendis. Optio et fugit perferendis aut placeat. Exercitationem expedita ipsam placeat quia eaque quaerat iure aut."
                                                        },
                                                        "cover_image_thumbnail_url": {
                                                            "type": "string",
                                                            "example": "https://via.placeholder.com/150/09f/fff.png"
                                                        },
                                                        "views_count": {
                                                            "type": "string",
                                                            "example": "1"
                                                        },
                                                        "likes_count": {
                                                            "type": "string",
                                                            "example": "1"
                                                        },
                                                        "comments": {
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "subject": {
                                                                        "type": "string",
                                                                        "example": "hjfjt745gjfg"
                                                                    },
                                                                    "comment": {
                                                                        "type": "string",
                                                                        "example": "iksjwkwswims"
                                                                    },
                                                                    "created_at": {
                                                                        "type": "string",
                                                                        "example": "2022-01-01 00:00:00"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "tags": {
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "": {
                                                                        "type": "string",
                                                                        "example": "technology"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Article data retrieved successfully."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/article/1/comment": {
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Comment on an article",
                "description": "comment on an article",
                "operationId": "Comment on an article",
                "requestBody": {
                    "description": "Comment on an article",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "body",
                                    "subject"
                                ],
                                "properties": {
                                    "subject": {
                                        "type": "string",
                                        "example": "jsljslkls"
                                    },
                                    "body": {
                                        "type": "text",
                                        "example": "Culpa facere laborum asperiores reiciendis. Optio et fugit perferendis aut placeat. Exercitationem expedita ipsam placeat quia eaque quaerat iure aut."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Returns parameters are not vaid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "subject is required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Returns if article is not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Article not found"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "data": {
                                            "properties": {
                                                "error": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": [
                                                            "Article not found"
                                                        ]
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "success": {
                                                    "type": "boolean",
                                                    "example": "true"
                                                },
                                                "data": {
                                                    "properties": {
                                                        "subject": {
                                                            "type": "string",
                                                            "example": "jsljslkls"
                                                        },
                                                        "comment": {
                                                            "type": "string",
                                                            "example": "Culpa facere laborum asperiores reiciendis. Optio et fugit perferendis aut placeat. Exercitationem expedita ipsam placeat quia eaque quaerat iure aut."
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "example": "2022-10-23T01:23:34.211613Z"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Your comment has been successfully sent"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/article/1/like": {
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Like an article",
                "description": "Like an article",
                "operationId": "Like an article",
                "requestBody": {
                    "description": "Like an article",
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "Returns if article is not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Article not found"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "data": {
                                            "properties": {
                                                "error": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": [
                                                            "Article not found"
                                                        ]
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "success": {
                                                    "type": "boolean",
                                                    "example": "true"
                                                },
                                                "data": {
                                                    "properties": {
                                                        "likes_count": {
                                                            "type": "integer",
                                                            "example": 4
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Your comment has been successfully sent"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}